Move MarkdownPage to page types directory
