+++ components/PageTypes/MarkdownPage.js
@@ -4,18 +4,15 @@ import ReactMarkdown from 'react-markdown';
 import Error from 'next/error';
 import 'styles/static-page.scss';
 
-const MarkdownPage = props => {
-  const {
-    data, error, pageTitle, pageSubTitle, children
-  } = props;
-
+const MarkdownPage = ( { data, error, pageTitle, pageSubTitle, children } ) => {
   if ( error ) {
     return <Error statusCode={ error } />;
   }
 
   return (
     <div className="static_page">
-      <h1>{ pageTitle }
+      <h1>
+        { pageTitle }
         { pageSubTitle && <div className="sub header">{ pageSubTitle }</div> }
       </h1>
       { children }
@@ -29,7 +26,7 @@ MarkdownPage.propTypes = {
   pageSubTitle: PropTypes.string,
   children: PropTypes.element,
   data: PropTypes.string,
-  error: PropTypes.oneOfType( [PropTypes.number, PropTypes.bool] )
+  error: PropTypes.oneOfType( [PropTypes.number, PropTypes.bool] ),
 };
 
 export default MarkdownPage;
