+++ pages/documentation/search.js
@@ -3,27 +3,23 @@ import PropTypes from 'prop-types';
 import config from 'config';
 
 import DocumentationMenu from 'components/DocumentationMenu/DocumentationMenu';
-import MarkdownPage from 'components/MarkdownPage';
+import MarkdownPage from 'components/PageTypes/MarkdownPage';
 import withCachedFetch from 'hocs/withCachedFetch/withCachedFetch';
 
-const Search = props => {
-  const { data, error } = props;
-
-  return (
-    <MarkdownPage
-      pageTitle="Content Commons Documentation"
-      pageSubTitle="Click on the links below for how-to guides on each subject. Please use Google Chrome for the best experience."
-      data={ data }
-      error={ error }
-    >
-      <DocumentationMenu />
-    </MarkdownPage>
-  );
-};
+const Search = ( { data, error } ) => (
+  <MarkdownPage
+    pageTitle="Content Commons Documentation"
+    pageSubTitle="Click on the links below for how-to guides on each subject. Please use Google Chrome for the best experience."
+    data={ data }
+    error={ error }
+  >
+    <DocumentationMenu />
+  </MarkdownPage>
+);
 
 Search.propTypes = {
   data: PropTypes.string,
-  error: PropTypes.oneOfType( [PropTypes.number, PropTypes.bool] )
+  error: PropTypes.oneOfType( [PropTypes.number, PropTypes.bool] ),
 };
 
 export default withCachedFetch( Search, config.SEARCHINGDOC_URL );
